+++ templates/ajax-overlay.php
@@ -6,7 +6,7 @@
  *
  * Page: "Jobs"
  *
- * @version 1.2.6
+ * @version 1.0.0
  */
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
