+++ templates/emails/base_wrapper.php
@@ -1,14 +1,13 @@
-<?php
+<?php // phpcs:ignore WordPress.Files.FileName.NotHyphenatedLowercase -- todo changing email notifications keys
 /**
  * Template for the base wrapper email template
  *
  * This template can be overridden by copying it to yourtheme/jobboardwp/emails/base_wrapper.php
  *
- * @version 1.2.6
+ * @version 1.1.0
  *
  * @var array $jb_emails_base_wrapper
  */
-// phpcs:ignore WordPress.Files.FileName.NotHyphenatedLowercase -- todo changing email notifications keys
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
