+++ templates/job/notices.php
@@ -4,7 +4,7 @@
  *
  * This template can be overridden by copying it to yourtheme/jobboardwp/job/notices.php
  *
- * @version 1.2.6
+ * @version 1.1.0
  *
  * @var array $jb_job_notices
  */
