+++ templates/jobs/list.php
@@ -4,7 +4,7 @@
  *
  * This template can be overridden by copying it to yourtheme/jobboardwp/jobs/list.php
  *
- * @version 1.2.6
+ * @version 1.2.0
  *
  * @var array $jb_jobs_list
  */
